refactor and import cleanup
