+++ test/tests/test_spectrum.py
@@ -5,9 +5,9 @@ from typing import Type
 
 import numpy as np
 from numpy.typing import NDArray
-
 import pytest
 
+import ramannoodle.io as rn_io
 from ramannoodle.polarizability.interpolation import InterpolationModel
 from ramannoodle.polarizability.art import ARTModel
 from ramannoodle.spectrum.raman import (
@@ -15,7 +15,6 @@ from ramannoodle.spectrum.raman import (
     get_laser_correction,
 )
 from ramannoodle.symmetry.structural import ReferenceStructure
-from ramannoodle import io
 from ramannoodle.spectrum.spectrum_utils import convolve_spectrum
 
 # pylint: disable=protected-access,too-many-locals
@@ -33,7 +32,7 @@ def _validate_polarizabilities(model: InterpolationModel, data_directory: str) -
     This function will use all *eps_OUTCAR's in a directory as references.
     """
     for outcar_path in _get_all_eps_outcars(data_directory):
-        positions, known_polarizability = io.read_positions_and_polarizability(
+        positions, known_polarizability = rn_io.read_positions_and_polarizability(
             f"{outcar_path}", file_format="outcar"
         )
         cartesian_displacement = model._ref_structure.get_cartesian_displacement(
@@ -73,7 +72,7 @@ def test_interpolation_spectrum(
     """Test a full spectrum calculation using InterpolationModel."""
     # Setup model
     ref_structure = outcar_ref_structure_fixture
-    _, polarizability = io.read_positions_and_polarizability(
+    _, polarizability = rn_io.read_positions_and_polarizability(
         f"{data_directory}/ref_eps_OUTCAR", file_format="outcar"
     )
     model = InterpolationModel(ref_structure, polarizability)
@@ -88,7 +87,7 @@ def test_interpolation_spectrum(
 
     # Spectrum test
     with np.load(f"{data_directory}/known_spectrum.npz") as known_spectrum:
-        phonons = io.read_phonons(
+        phonons = rn_io.read_phonons(
             f"{data_directory}/phonons_OUTCAR", file_format="outcar"
         )
         spectrum = phonons.get_raman_spectrum(model)
@@ -134,7 +133,7 @@ def test_art_spectrum(
     """Test a full spectrum calculation using ARTModel."""
     # Setup model
     ref_structure = outcar_ref_structure_fixture
-    _, polarizability = io.read_positions_and_polarizability(
+    _, polarizability = rn_io.read_positions_and_polarizability(
         f"{data_directory}/ref_eps_OUTCAR", file_format="outcar"
     )
     model = ARTModel(ref_structure, polarizability)
@@ -145,7 +144,7 @@ def test_art_spectrum(
 
     # Spectrum test
     with np.load(f"{data_directory}/known_art_spectrum.npz") as known_spectrum:
-        phonons = io.read_phonons(
+        phonons = rn_io.read_phonons(
             f"{data_directory}/phonons_OUTCAR", file_format="outcar"
         )
         spectrum = phonons.get_raman_spectrum(model)
@@ -212,7 +211,7 @@ def test_art_masked_spectrum(
     """Test a masked spectrum calculation using ARTModel."""
     # Setup model
     ref_structure = outcar_ref_structure_fixture
-    _, polarizability = io.read_positions_and_polarizability(
+    _, polarizability = rn_io.read_positions_and_polarizability(
         f"{data_directory}/ref_eps_OUTCAR", file_format="outcar"
     )
     model = ARTModel(ref_structure, polarizability)
@@ -225,7 +224,7 @@ def test_art_masked_spectrum(
 
     # Spectrum test
     with np.load(f"{data_directory}/{known_spectrum_file}") as known_spectrum:
-        phonons = io.read_phonons(
+        phonons = rn_io.read_phonons(
             f"{data_directory}/phonons_OUTCAR", file_format="outcar"
         )
         spectrum = phonons.get_raman_spectrum(model)
